[Feat] do not persist key if already exists

+++ backend/src/service/dataService.ts
@@ -36,8 +36,10 @@ export enum Key {
     Private = "privateKey"
 }
 
-export const saveKey = (key: Uint8Array, keyName: Key) => {
-    fs.writeFileSync(config.baseDir + 'user/' + keyName, Buffer.from(key));
+export const saveKey = (key: Uint8Array, keyName: Key, force = false) => {
+    if(force || !fs.existsSync(config.baseDir + 'user/' + keyName)) {
+        fs.writeFileSync(config.baseDir + 'user/' + keyName, Buffer.from(key));
+    }
 }
 
 export const getKey = (keyName: string): Uint8Array | undefined => {
